Basic item logic
The basics needed to have a inventory and item system, and some visual structures.

+++ Assets/Scripts/Player/PlayerStatusUIManager.cs
@@ -37,7 +37,7 @@ public class PlayerStatusUIManager : MonoBehaviour{
         }
     }
 
-    public void ChangeSliderHealth(int value){
+    public void ChangeSliderHealth(float value){
         healthSlider.value = value;
 
         healthFillImage.color = healthGradient.Evaluate(healthSlider.normalizedValue);
